+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-## Threads
-- Added another method to create a detached thread
+## General
+- Changed custom array element at macro definition name
+
+## HTTP
+- Changed HTTP invalid header type definition to undefined
+- Added dedicated custom HTTP header structure definition
+- Changed temp HTTP header size actual value in sources
+- Added the ability for requests to handle custom headers
 
 ## Tests
-- Added dedicated thread_create_detached () unit test
+- Updated headers unit tests methods with latest definitions
+- Added HTTP request custom headers methods unit tests